// packages/api/src/generated/user/models/listPublicConcernsSort.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Ogonggo User API
 * OpenAPI spec version: v1
 */

export type ListPublicConcernsSort = typeof ListPublicConcernsSort[keyof typeof ListPublicConcernsSort];


export const ListPublicConcernsSort = {
  LATEST: 'LATEST',
  VIEW_COUNT: 'VIEW_COUNT',
  COMMENT_COUNT: 'COMMENT_COUNT',
} as const;