// packages/api/src/generated/user/models/listPublicPopularConcernsSort.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Ogonggo User API
 * OpenAPI spec version: v1
 */

export type ListPublicPopularConcernsSort = typeof ListPublicPopularConcernsSort[keyof typeof ListPublicPopularConcernsSort];


export const ListPublicPopularConcernsSort = {
  VIEW_COUNT: 'VIEW_COUNT',
  COMMENT_COUNT: 'COMMENT_COUNT',
} as const;