// packages/api/src/generated/user/models/createConcernCommentRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Ogonggo User API
 * OpenAPI spec version: v1
 */

export interface CreateConcernCommentRequest {
  /**
     * @minLength 0
     * @maxLength 1000
     */
  content: string;
  /** 답글을 달 답변(부모 댓글) 식별자. 보내지 않으면 답변을 작성합니다. */
  parentId?: number;
}