// packages/api/src/generated/user/models/saveConcernRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Ogonggo User API
 * OpenAPI spec version: v1
 */
import type { SaveConcernRequestCategory } from './saveConcernRequestCategory';

export interface SaveConcernRequest {
  /**
     * | 값 | code | 설명 |
     * | --- | --- | --- |
     * | `JOB_POSTING` | 1 | 공고 질문 |
     * | `CAREER` | 2 | 직무·커리어 |
     * | `APPLICATION_INTERVIEW` | 3 | 서류·면접 |
     * | `SIDE_EXPERIENCE` | 4 | 사이드·경험 |
     * | `ETC` | 5 | 기타 |
     */
  category: SaveConcernRequestCategory;
  /**
     * @minLength 0
     * @maxLength 100
     */
  title: string;
  /**
     * @minLength 0
     * @maxLength 2000
     */
  content: string;
}