// packages/api/src/generated/user/models/successResponseConcernDetailResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Ogonggo User API
 * OpenAPI spec version: v1
 */
import type { ConcernDetailResponse } from './concernDetailResponse';

export interface SuccessResponseConcernDetailResponse {
  status: number;
  message: string;
  data?: ConcernDetailResponse;
}