// packages/api/src/generated/user/models/successResponseCreateConcernCommentResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Ogonggo User API
 * OpenAPI spec version: v1
 */
import type { CreateConcernCommentResponse } from './createConcernCommentResponse';

export interface SuccessResponseCreateConcernCommentResponse {
  status: number;
  message: string;
  data?: CreateConcernCommentResponse;
}