// packages/api/src/generated/user/models/successResponseCreateConcernResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Ogonggo User API
 * OpenAPI spec version: v1
 */
import type { CreateConcernResponse } from './createConcernResponse';

export interface SuccessResponseCreateConcernResponse {
  status: number;
  message: string;
  data?: CreateConcernResponse;
}